Ajout API  recherche avatar + recherche avatar avec descriptif

+++ SportSocial/src/main/java/fr/solutec/entities/Produit.java
@@ -17,7 +17,7 @@ public class Produit {
 	private Long idProduit;
 	
 	private String descriptionProduit;
-	private Float prixTokenProduit;
+	private Double prixTokenProduit;
 	@ManyToOne
 	private Image imageProduit;
 }
